Drum_midi/requirements.txt
mido
python-rtmidi
numpy
simpleaudio
pyserial
